Revert "Plan 9 from Bell Labs Fourth Edition 2006-12-23"
This reverts commit c93608cc76758b2be624199c6208a0f90bad298d.

+++ sys/src/cmd/db/command.c
@@ -18,7 +18,7 @@ char	stformat[ARB] = "zMi";
 ADDR	ditto;
 
 ADDR	dot;
-int	dotinc;
+WORD	dotinc;
 WORD	adrval, cntval, loopcnt;
 int	adrflg, cntflg;
